// src/lib/temple/types.ts
export interface TempleNetwork {
  id: string;
  name: string;
  nameI18nKey?: string;
  description: string;
  descriptionI18nKey?: string;
  type: TempleNetworkType;
  rpcBaseURL: string;
  color: string;
  disabled: boolean;
  hidden?: boolean;
}

export type TempleNetworkType = 'main' | 'test';